workout

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,19 @@
 import { useContext } from "react";
 import { WorkoutContext } from "../context/WorkoutContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 function WorkoutDetails({ workout }) {
   const { dispatch } = useContext(WorkoutContext);
+  const { user } = useAuthContext();
 
   const handleCancel = async () => {
+    if (!user) return;
+
     const res = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const data = await res.json();
